backend/requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.3.0
audioread==3.0.1
babel==2.16.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==25.1.0
bleach==6.2.0
blinker==1.8.2
capstone==6.0.0a2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
colored-traceback==0.4.2
comm==0.2.0
contourpy==1.2.0
cookiecutter==2.6.0
cryptography==45.0.3
cycler==0.12.1
databases==0.9.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
executing==2.0.1
Faker==37.3.0
fastapi==0.115.12
fastapi-cli==0.0.6
fastjsonschema==2.21.1
Flask==3.0.3
flet==0.28.3
flet-cli==0.28.3
flet-desktop==0.28.3
flet-web==0.28.3
fonttools==4.44.0
fqdn==1.5.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
intervaltree==3.1.0
ipykernel==6.26.0
ipympl==0.9.6
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mido==1.3.3
mistune==3.1.0
msgpack==1.1.0
mypy_extensions==1.1.0
mysql-connector==2.2.9
mysql-connector-python==9.3.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.5.8
notebook==7.3.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.65.2
overrides==7.7.0
packaging==25.0
pandas==2.1.2
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.3
pathspec==0.12.1
pdfminer.six==20250327
pillow==11.2.1
platformdirs==4.3.6
pluggy==1.6.0
plumbum==1.9.0
pooch==1.8.2
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psutil==5.9.6
pure-eval==0.2.2
pwntools==4.13.1
pyahocorasick==2.2.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyelftools==0.31
Pygments==2.19.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==3.1.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pypng==0.20220715.0
PyQRCode==1.2.1
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
pyqt5-tools==5.15.9.3.3
PyQt5_sip==12.15.0
pyserial==3.5
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.2.1
python-multipart==0.0.19
python-slugify==8.0.4
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==25.1.1
qrcode==7.4.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
rarfile==4.2
referencing==0.36.1
repath==0.9.0
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-toolkit==0.12.0
ROPGadget==7.5
rpds-py==0.22.3
rpyc==6.0.1
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.0
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.3.3
tqdm==4.67.1
traitlets==5.13.0
typer==0.15.1
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
unicorn==2.1.1
unix-ar==0.2.1
uri-template==1.3.0
urllib3==2.4.0
uv==0.7.7
uvicorn==0.34.3
watchdog==4.0.2
watchfiles==1.0.5
wcwidth==0.2.9
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.4
widgetsnbextension==4.0.13
zstandard==0.23.0
